# docs/smmp-api.openapi.yml
openapi: 3.0.0
info:
  title: SM Marketplace API
  version: 1.0.0
servers:
  - url: http://{{host}}
paths:
  /upload-file:
    post:
      tags:
        - default
      summary: Upload file
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  description: archivo a subir
                  format: binary
                metadata:
                  type: string
                  description: los valores deben ser numeros o string
                  example: '{"subject":"test"}'
      responses:
        '200':
          description: OK
          headers:
            X-Powered-By:
              schema:
                type: string
                example: Express
            Access-Control-Allow-Origin:
              schema:
                type: string
                example: '*'
            Content-Type:
              schema:
                type: string
                example: application/json; charset=utf-8
            Content-Length:
              schema:
                type: integer
                example: '271'
            ETag:
              schema:
                type: string
                example: W/"10f-8nGP4qbgEQNQxAvd14ZjKZGqlOI"
            Date:
              schema:
                type: string
                example: Fri, 02 Dec 2022 07:24:26 GMT
            Connection:
              schema:
                type: string
                example: keep-alive
            Keep-Alive:
              schema:
                type: string
                example: timeout=5
          content:
            application/json:
              schema:
                type: object
              example:
                status: true
                message: File is uploaded
                data:
                  name: rintaro-1.jpg
                  mimetype: image/jpeg
                  size: 194547
                pinata:
                  IpfsHash: QmYSunLGem8Jqpr4ckDZjQhEywUya3T7Lxhpk1XbNMEg3j
                  PinSize: 194561
                  Timestamp: '2022-12-02T07:24:26.275Z'
                metadata:
                  subject: test
  /search:
    post:
      tags:
        - default
      summary: Pinata Search
      requestBody:
        content:
          application/json:
            schema:
              type: object
              example:
                filters:
                  hashContains: QmRZVAozYgkw3C8zyGAaBQGE81zBV5ekFP7BiTc2Tfw3Gy
      responses:
        '200':
          description: OK
          headers:
            X-Powered-By:
              schema:
                type: string
                example: Express
            Access-Control-Allow-Origin:
              schema:
                type: string
                example: '*'
            Content-Type:
              schema:
                type: string
                example: application/json; charset=utf-8
            Content-Length:
              schema:
                type: integer
                example: '511'
            ETag:
              schema:
                type: string
                example: W/"1ff-6bT3FCuSjbE3GCXWGwiaafycUmk"
            Date:
              schema:
                type: string
                example: Fri, 02 Dec 2022 06:44:42 GMT
            Connection:
              schema:
                type: string
                example: keep-alive
            Keep-Alive:
              schema:
                type: string
                example: timeout=5
          content:
            application/json:
              schema:
                type: object
              example:
                status: true
                itemsFound: 1
                items:
                  - id: 523346c4-2070-4234-a5af-6aaacf065d5f
                    ipfs_pin_hash: QmRZVAozYgkw3C8zyGAaBQGE81zBV5ekFP7BiTc2Tfw3Gy
                    size: 155841
                    user_id: c62b3ae7-7f61-4633-a8f8-59fc2224b358
                    date_pinned: '2022-11-25T08:40:13.972Z'
                    date_unpinned: null
                    metadata:
                      name: mayuri.jpg
                      keyvalues:
                        uploadedBy: SMMP API
                    regions:
                      - regionId: FRA1
                        currentReplicationCount: 1
                        desiredReplicationCount: 1
                      - regionId: NYC1
                        currentReplicationCount: 1
                        desiredReplicationCount: 1